Split Commands.py into Fun.py and Utils.py
